# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI TimeShards API",
    "version": "0.1.0",
    "description": "REST API for time tracking and access control. Narrative reference: docs/API.md. Machine-readable spec served at GET /api/v1/openapi.json."
  },
  "servers": [{ "url": "http://127.0.0.1:47821", "description": "Default local bind" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token from POST /api/v1/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/reports/timesheets/export": {
      "get": {
        "summary": "Export timesheets (CSV or HTML)",
        "parameters": [
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["csv", "html"], "default": "csv" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "period_start", "in": "query", "schema": { "type": "string", "format": "date-time" } }
        ]
      }
    },
    "/api/v1/reports/access/export": {
      "get": {
        "summary": "Export access log (CSV or HTML)",
        "parameters": [
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["csv", "html"], "default": "csv" } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "maximum": 2000 } }
        ]
      }
    }
  },
  "security": [{ "bearerAuth": [] }]
}